Move pruning utils to pruning_handler and remove time basis pruning
