+++ bftengine/include/bftengine/ReplicaConfig.hpp
@@ -112,7 +112,6 @@ class ReplicaConfig : public concord::serialize::SerializableFactory<ReplicaConf
   // Pruning parameters
   CONFIG_PARAM(pruningEnabled_, bool, false, "Enable pruning");
   CONFIG_PARAM(numBlocksToKeep_, uint64_t, 0, "how much blocks to keep while pruning");
-  CONFIG_PARAM(durationToKeppMinutes_, uint64_t, 0, "how old the data to prune");
 
   CONFIG_PARAM(debugPersistentStorageEnabled, bool, false, "whether persistent storage debugging is enabled");
 
@@ -191,7 +190,6 @@ class ReplicaConfig : public concord::serialize::SerializableFactory<ReplicaConf
 
     serialize(outStream, pruningEnabled_);
     serialize(outStream, numBlocksToKeep_);
-    serialize(outStream, durationToKeppMinutes_);
 
     serialize(outStream, debugPersistentStorageEnabled);
     serialize(outStream, maxExternalMessageSize);
@@ -244,7 +242,6 @@ class ReplicaConfig : public concord::serialize::SerializableFactory<ReplicaConf
 
     deserialize(inStream, pruningEnabled_);
     deserialize(inStream, numBlocksToKeep_);
-    deserialize(inStream, durationToKeppMinutes_);
 
     deserialize(inStream, debugPersistentStorageEnabled);
     deserialize(inStream, maxExternalMessageSize);
